feat(graph): Add coordinates to nodes, remove Kruskal as a class and add Glouton algorithm

+++ src/main/java/dev/e_psi_lon/Arc.java
@@ -1,6 +1,6 @@
 package dev.e_psi_lon;
 
-public record Arc(Noeud source, Noeud cible, int weight) {
+public record Arc(Noeud source, Noeud cible, double weight) {
     public Arc(Noeud source, Noeud cible) {
         this(source, cible, 1);
     }
